// libs/shared/src/types/misc/core/metric-type.enum.ts
export enum METRIC_TYPE {
  WEIGHT = 'WEIGHT',
  HEIGHT = 'HEIGHT',
  BMI = 'BMI',
  BODY_FAT = 'BODY_FAT',
  MUSCLE_MASS = 'MUSCLE_MASS',
  DAILY_CALORIES = 'DAILY_CALORIES',
  DAILY_STEPS = 'DAILY_STEPS',
  SLEEP_DURATION = 'SLEEP_DURATION',
  SLEEP_SCORE = 'SLEEP_SCORE',
  HOOPER_INDEX = 'HOOPER_INDEX',
  HR_MAX = 'HR_MAX',
  HR_REST = 'HR_REST',
  HR_RESERVE = 'HR_RESERVE',
  RMSSD = 'RMSSD',
  VMA = 'VMA',
  CRITICAL_POWER_RUNNING = 'CRITICAL_POWER_RUNNING',
  CRITICAL_POWER_CYCLING = 'CRITICAL_POWER_CYCLING',
  VO2MAX = 'VO2MAX',
  FTP_RUNNING = 'FTP_RUNNING',
  FTP_CYCLING = 'FTP_CYCLING',
  VERTICAL_SPEED_AVG = 'VERTICAL_SPEED_AVG',
  VERTICAL_SPEED_MAX = 'VERTICAL_SPEED_MAX',
  FITNESS_INDEX = 'FITNESS_INDEX',
}